requirements.txt
streamlit
python-dotenv
pandas
sentence-transformers
fastembed
qdrant-client
langchain-google-genai
langchain-core
torch
numpy
tqdm
